safari-600.7-branch/LayoutTests/fast/dom/compare-document-position-disconnected-nodes-expected.txt

PASS a.compareDocumentPosition(b) & Node.DOCUMENT_POSITION_DISCONNECTED is Node.DOCUMENT_POSITION_DISCONNECTED
PASS b.compareDocumentPosition(a) & Node.DOCUMENT_POSITION_DISCONNECTED is Node.DOCUMENT_POSITION_DISCONNECTED
PASS a.compareDocumentPosition(b) & Node.DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC is Node.DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC
PASS b.compareDocumentPosition(a) & Node.DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC is Node.DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC
PASS a.compareDocumentPosition(b) & Node.DOCUMENT_POSITION_PRECEDING is 0
PASS a.compareDocumentPosition(b) & Node.DOCUMENT_POSITION_FOLLOWING is 0
PASS successfullyParsed is true

TEST COMPLETE
Comparing the positions of two disconnected nodes should result in DOCUMENT_POSITION_DISCONNECTED.
